requirements.txt
Pillow==8.1.0
requests==2.25.1
pytesseract==0.3.7
python-telegram-bot==13.2
torch==1.7.1
torchvision==0.8.2